clicker

+++ Assets/App/Core/Menu/External/Animations/ISoftAccrualAnimation.cs
@@ -4,6 +4,6 @@ namespace App.Menu.UI.External.Animations
 {
     public interface ISoftAccrualAnimation
     {
-        void PlayAnimation(Vector3 globalPosition, Transform parent, int amount);
+        void PlayAnimation(Vector3 globalPosition, Transform parent, long amount);
     }
 }
